use separate file to avoid making qgis.h bigger
