+++ src/providers/arcgisrest/qgsnewarcgisrestconnection.cpp
@@ -19,6 +19,7 @@
 #include "qgssettings.h"
 #include "qgshelp.h"
 #include "qgsgui.h"
+#include "fromencodedcomponenthelper.h"
 
 #include <QMessageBox>
 #include <QUrl>
@@ -160,7 +161,7 @@ QUrl QgsNewArcGisRestConnectionDialog::urlTrimmed() const
 
   if ( url.path( QUrl::FullyEncoded ).isEmpty() )
   {
-    url.setPath( qgis::fromEncodedComponent_helper( "/" ) );
+    url.setPath( fromEncodedComponent_helper( "/" ) );
   }
   return url;
 }
